requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
asyncpg
cryptography==42.0.5
python-dotenv==1.0.1
openpyxl==3.1.2
